SLIM-882 Implements GenAndReplaceKeys for an action bundle
